requirements.txt
pandas
numpy
requests
sqlalchemy
psycopg2-binary
python-dotenv
yfinance
streamlit
plotly
